requirements.txt
asgiref==3.7.2
click==8.1.7
contourpy==1.2.1
crispy-bootstrap4==2024.1
cycler==0.12.1
distlib==0.3.8
Django==4.2.7
django-crispy-forms==2.1
filelock==3.13.4
fonttools==4.51.0
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.8.2
nltk==3.8.1
numpy==1.26.3
packaging==24.0
pandas==2.2.0
pillow==10.3.0
platformdirs==4.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3.post1
regex==2024.4.16
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.4.0
tqdm==4.66.2
tzdata==2024.1
virtualenv==20.25.3
